Add ForeignKey relation and decorators

+++ main/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from .forms import JssForm
 from .models import Jasoseol
+from django.core.exceptions import PermissionDenied
+from django.contrib.auth.decorators import login_required
 # from django.http import Http404
 
 # Create your views here.
@@ -9,15 +11,26 @@ def index(request):
     all_jss = Jasoseol.objects.all()
     return render(request, 'index.html', {'all_jss' : all_jss})
 
+def my_index(request):
+    my_jss = Jasoseol.objects.filter(author=request.user)
+    return render(request, 'index.html', {'all_jss' : my_jss})
+
+@login_required(login_url='/login/')
 def create(request):
+    # if not request.user.is_authenticated:
+    #     return redirect('login')
     if request.method == 'POST':
         filled_form = JssForm(request.POST)
         if filled_form.is_valid():
-            filled_form.save()
+            temp_form = filled_form.save(commit=False)
+            # commit=False : object 저장을 지연시키고 추가적인 동작이 가능
+            temp_form.author = request.user
+            temp_form.save()
             return redirect('index')
     jss_form = JssForm()
     return render(request, 'create.html', {'jss_form' : jss_form})
 
+@login_required(login_url='/login/')
 def detail(request, jss_id):
     # try:
     #     my_jss = Jasoseol.objects.get(pk=jss_id)
@@ -28,15 +41,21 @@ def detail(request, jss_id):
 
 def delete(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
-    my_jss.delete()
-    return redirect('index')
+    if request.user == my_jss.author:
+        my_jss.delete()
+        return redirect('index')
+    
+    raise PermissionDenied
 
 def update(request, jss_id):
     my_jss = Jasoseol.objects.get(pk=jss_id)
-    jss_form = JssForm(instance=my_jss)
-    if request.method == 'POST':
-        updated_form =  JssForm(request.POST, instance=my_jss)
-        if updated_form.is_valid():
-            updated_form.save()
-            return redirect('index')
-    return render(request, 'create.html', {'jss_form':jss_form})
+    if request.user == my_jss.author:
+        jss_form = JssForm(instance=my_jss)
+        if request.method == 'POST':
+            updated_form =  JssForm(request.POST, instance=my_jss)
+            if updated_form.is_valid():
+                updated_form.save()
+                return redirect('index')
+        return render(request, 'create.html', {'jss_form':jss_form})
+
+    raise PermissionDenied
